Add game state management and enhance input handling: implement GameState enum, update MouseHandler and KeyboardHandler for state-based actions, and improve hitbox initialization

+++ src/main/java/org/unibo/entities/Entity.java
@@ -18,8 +18,8 @@ public abstract class Entity {
         this.height = height;
     }
 
-    protected void initHitBox(float x, float y, float width, float height) {
-        hitBox = new Rectangle2D.Float((int)x, (int)y, width, height);
+    protected void initHitBox(float x, float y, int width, int height) {
+        hitBox = new Rectangle2D.Float(x, y, width, height);
     }
 
     public Rectangle2D.Float getHitBox() {
